Refactor: extract search date helpers into _dates.py
Move _DATE_ONLY_FIELDS, keyword constants, _DATE_KEYWORDS,
_DATE_KEYWORD_PATTERN, _fmt, _iso_range, _date_only_range, and
_datetime_range from _query.py into a new dependency-free _dates.py.
Re-export the symbols used by _query.py for backward compatibility.
No behavior change; all 100 search query tests pass.
